--- pkg/util/filter_test.go
@@ -1,52 +1,0 @@
-/*
-Copyright 2018 The Jetstack cert-manager contributors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package util
-
-import "testing"
-
-func TestRemoveDuplicates(t *testing.T) {
-	type testT struct {
-		input  []string
-		output []string
-	}
-	tests := []testT{
-		{
-			input:  []string{"a"},
-			output: []string{"a"},
-		},
-		{
-			input:  []string{"a", "b"},
-			output: []string{"a", "b"},
-		},
-		{
-			input:  []string{"a", "a"},
-			output: []string{"a"},
-		},
-		{
-			input:  []string{"a", "b", "a", "a", "c"},
-			output: []string{"a", "b", "c"},
-		},
-	}
-	for _, test := range tests {
-		actualOutput := RemoveDuplicates(test.input)
-		if len(actualOutput) != len(test.output) ||
-			!EqualUnsorted(test.output, actualOutput) {
-			t.Errorf("returned %q for %q but expected %q", actualOutput, test.input, test.output)
-			continue
-		}
-	}
-}
