+++ pkg/util/pki/csr_test.go
@@ -21,6 +21,7 @@ import (
 	"testing"
 
 	"github.com/jetstack/cert-manager/pkg/apis/certmanager/v1alpha1"
+	"github.com/jetstack/cert-manager/pkg/util"
 )
 
 func buildCertificate(cn string, dnsNames ...string) *v1alpha1.Certificate {
@@ -231,5 +232,38 @@ func TestSignatureAlgorithmForCertificate(t *testing.T) {
 
 	for _, test := range tests {
 		t.Run(test.name, testFn(test))
+	}
+}
+
+func TestRemoveDuplicates(t *testing.T) {
+	type testT struct {
+		input  []string
+		output []string
+	}
+	tests := []testT{
+		{
+			input:  []string{"a"},
+			output: []string{"a"},
+		},
+		{
+			input:  []string{"a", "b"},
+			output: []string{"a", "b"},
+		},
+		{
+			input:  []string{"a", "a"},
+			output: []string{"a"},
+		},
+		{
+			input:  []string{"a", "b", "a", "a", "c"},
+			output: []string{"a", "b", "c"},
+		},
+	}
+	for _, test := range tests {
+		actualOutput := removeDuplicates(test.input)
+		if len(actualOutput) != len(test.output) ||
+			!util.EqualUnsorted(test.output, actualOutput) {
+			t.Errorf("returned %q for %q but expected %q", actualOutput, test.input, test.output)
+			continue
+		}
 	}
 }
